+++ plugins/files/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @matatbread/matbot-files-node
 
+## 0.3.9
+
+### Patch Changes
+
+- @matatbread/matbot-core@0.3.9
+
 ## 0.3.8
 
 ### Patch Changes