+++ plugins/providers/google/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @matatbread/matbot-provider-google
 
+## 0.3.9
+
+### Patch Changes
+
+- @matatbread/matbot-core@0.3.9
+- @matatbread/matbot-plugin-api@0.3.9
+- @matatbread/matbot-provider-openai-compat@0.3.9
+
 ## 0.3.8
 
 ### Patch Changes