+++ plugins/skills-node/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @matatbread/matbot-skills-node
 
+## 0.3.9
+
+### Patch Changes
+
+- @matatbread/matbot-plugin-api@0.3.9
+- @matatbread/matbot-skills@0.3.9
+
 ## 0.3.8
 
 ### Patch Changes