+++ plugins/storage/google-drive/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @matatbread/matbot-storage-google-drive
 
+## 0.3.9
+
+### Patch Changes
+
+- Updated dependencies
+  - @matatbread/matbot-browser@0.3.9
+  - @matatbread/matbot-core@0.3.9
+  - @matatbread/matbot-plugin-api@0.3.9
+
 ## 0.3.8
 
 ### Patch Changes